Add files via upload
This dashboard was recreated using a Python script, where Matplotlib was utilized for certain figures that were challenging to replicate with Bokeh. Moving forward, I plan to verify the feasibility of creating all figures in Bokeh before beginning the development process. Sticking to a single library like Bokeh would be preferable due to its superior capabilities in dashboard construction and interactive features.